requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
requests==2.31.0
beautifulsoup4==4.12.3
google-generativeai==0.3.2
xhtml2pdf==0.2.13
PyMuPDF==1.23.8
Pillow==10.1.0
duckduckgo-search==3.9.6
playwright==1.40.0
python-dotenv==1.0.0
Werkzeug==3.0.1
